% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_crop.R
\name{gdal_crop}
\alias{gdal_crop}
\title{gdal_crop
Crop raster}
\usage{
gdal_crop(
  input_raster,
  filename,
  shapefile_path,
  large_tif = FALSE,
  return_raster = TRUE
)
}
\arguments{
\item{input_raster}{Character. Raster to be cropped (file path or Raster object). Input raster should be stored on disk for GDAL access.}

\item{filename}{Character. Output raster path.}

\item{shapefile_path}{Character. Filepath of a shapefile or geopackage. that is used to crop the raster. Disk path, \strong{not sp or sf object}.}

\item{large_tif}{Logical. Use \code{large_tif = TRUE} for large rasters (>4GB).}

\item{return_raster}{Logical. The function stores the raster in the \code{filename} argument path as a side-effect.}
}
\value{
Raster object. Only if \code{return_raster = TRUE}. Otherwise, the function side-effect is to save the file locally.
}
\description{
The function crops rasters (to the shapefile boundary) using GDAL. Use only filepaths (not rasters/shapefiles loaded in R environment).
}
\examples{
library(gdalR)
library(Rahat)

we_eu <- milkunize("Western_Europe_IMAGE.shp")
dem <- milkunize("Merit_DEM_10s.tif", "m5")

# Crop dem
dem_cropped <- GDAL_crop(dem, filename = "/vol/milkun1/Mirza_Cengic/Temp/dem_we.tif", shapefile_path = we_eu)
}
