% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_mosaic_tile.R
\name{gdal_mosaic_tile}
\alias{gdal_mosaic_tile}
\title{gdal_mosaic_tile
Mosaic raster tiles}
\usage{
gdal_mosaic_tile(
  output_file,
  folder_path,
  return_raster = TRUE,
  large_tif = TRUE
)
}
\arguments{
\item{output_file}{Path and filename of the mosaicked output (format "E:/Folder/file.tif")}

\item{folder_path}{Folder path with the tif files to be mosaiced (format "E:/Folder/*.tif")}

\item{return_raster}{Return raster.}

\item{large_tif}{Use \code{large_tif = TRUE} for large rasters (>4GB)}
}
\value{
raster.
}
\description{
gdal_mosaic_tile
Mosaic raster tiles
}
