% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_to_polygon.R
\name{gdal_raster_to_polygon}
\alias{gdal_raster_to_polygon}
\title{gdal_raster_to_polygon
Convert raster to polygon}
\usage{
gdal_raster_to_polygon(input_raster, output_file, return_shp = FALSE)
}
\arguments{
\item{input_raster}{Raster* file or path of raster file on disk}

\item{output_file}{Path of new shapefile}

\item{return_shapefile}{Logical. If TRUE, function will return a simple feature. Default is FALSE.}
}
\description{
gdal_raster_to_polygon
Convert raster to polygon
}
