% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDAL_resample.R
\name{GDAL_resample}
\alias{GDAL_resample}
\title{GDAL_resample}
\usage{
GDAL_resample(
  infile,
  outfile,
  resolutions,
  target_extent,
  method,
  large_tif = FALSE,
  return_raster = FALSE
)
}
\arguments{
\item{infile}{character  Raster to be cropped (file path or Raster object). Input raster should be stored on disk for GDAL access.}

\item{outfile}{character Path and filename of the mosaicked output (format "E:/Folder/file.tif").}

\item{resolutions}{numeric ResX e ResY as vector.}

\item{target_extent}{character New extent as xmin ymin xmax ymax".}

\item{method}{character Resampling method. One in c("near", "bilinear", "cubic", "cubicspline", "lanczos",
"average", "mode", "max", "min", "med", "q1", "q3").}

\item{large_tif}{logical  Use \code{large_tif = TRUE} for large rasters (>4GB).}

\item{return_raster}{logical The function stores the raster in the \code{filename} argument path as a side-effect.}
}
\value{
Raster object. Only if \code{return_raster = TRUE}. Otherwise, the function side-effect is to save the file locally.
}
\description{
Function does a system call to run gdal functions (works on unix systems where gdal can be called from terminal).
Use only file path, function will not work with raster objects loaded in R environment.
}
