% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDAL_get_changes.R
\name{get_changes}
\alias{get_changes}
\title{Extract changed cells from two rasters
Get changes from two rasters from different periods. The function is suitable for parallel processing of very large rasters on computers with many cores.}
\usage{
get_changes(x, y, size, outpath, outfile, values_of_interest, spatial = FALSE,
  number_of_cores = 2)
}
\arguments{
\item{x}{t1 raster}

\item{y}{t2 raster}

\item{size}{tilesize}

\item{outpath}{Folder in which the temporary raster files will be created. If keep_temp argument is set to FALSE (default), files will be deleted after the process is sucessfully executed. Default path is R temporary directory (see tempdir()).}

\item{outfile}{Output file name. Optional.}

\item{spatial}{Logical. When TRUE function returns SpatialPointsDataFrame. Default is FALSE.}

\item{crop_category_to_model}{crop category to subset. Numeric. Accepts vectors.}
}
\value{
None.
}
\description{
Extract changed cells from two rasters
Get changes from two rasters from different periods. The function is suitable for parallel processing of very large rasters on computers with many cores.
}
\examples{
Dunno
}
