% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDAL_crop.R
\name{GDAL_crop}
\alias{GDAL_crop}
\title{Crop raster}
\usage{
GDAL_crop(input_raster, filename, shapefile_path, return_raster = TRUE)
}
\arguments{
\item{input_raster}{Raster to be cropped (file path or Raster object). Input raster should be stored on disk for GDAL access.}

\item{filename}{Output raster path.}

\item{shapefile_path}{Shapefile that is used to crop the raster. Disk path, \strong{not sp or sf object}.}

\item{return_raster}{Logical. The function stores the raster in the \code{filename} argument path as a side-effect.}
}
\value{
Raster object. Only if \code{return_raster = TRUE}. Otherwise, the function side-effect is to save the file locally.
}
\description{
The function crops rasters (to the shapefile boundary) using GDAL.
}
\examples{
library(gdalR)
library(Rahat)

we_eu <- milkunize("Projects/Land_use/Data/Shapefile/IMAGE_regions/Individual_regions/Western_Europe_IMAGE.shp")
dem <- milkunize("Merit_DEM_10s.tif", "m5")

# Crop dem
dem_cropped <- GDAL_crop(dem, filename = "/vol/milkun1/Mirza_Cengic/Temp/dem_we.tif", shapefile_path = we_eu)
}
