% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDAL_mosaic_tile.R
\name{GDAL_mosaic_tile}
\alias{GDAL_mosaic_tile}
\title{Mosaic raster tiles}
\usage{
GDAL_mosaic_tile(output_file, folder_path, return_raster = TRUE,
  large_tif = TRUE)
}
\arguments{
\item{output_file}{Path and filename of the mosaicked output (format "E:/Folder/file.tif")}

\item{folder_path}{Folder path with the tif files to be mosaiced (format "E:/Folder/*.tif")}

\item{return_raster}{Return raster.}

\item{large_tif}{Use \code{large_tif = TRUE} for large rasters (>4GB)}
}
\value{
Raster.
}
\description{
Mosaic raster tiles
}
\examples{
None.
}
