% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDAL_resample.R
\name{GDAL_resample}
\alias{GDAL_resample}
\title{Resample raster}
\usage{
GDAL_resample(infile, outfile, target_resolution, method,
  large_tif = FALSE, return_raster = FALSE)
}
\arguments{
\item{infile}{Raster to be cropped (file path or Raster object). Input raster should be stored on disk for GDAL access.}

\item{outfile}{Path and filename of the mosaicked output (format "E:/Folder/file.tif").}

\item{target_resolution}{New resolution of raster.}

\item{method}{Resampling method. One in c("near", "bilinear", "cubic", "cubicspline", "lanczos",
"average", "mode", "max", "min", "med", "q1", "q3").}

\item{large_tif}{Use \code{large_tif = TRUE} for large rasters (>4GB).}

\item{return_raster}{Logical. The function stores the raster in the \code{filename} argument path as a side-effect.}
}
\value{
Raster object. Only if \code{return_raster = TRUE}. Otherwise, the function side-effect is to save the file locally.
}
\description{
Function does a system call to run gdal functions (works on unix systems where gdal can be called from terminal).
Use only file path, function will not work with raster objects loaded in R environment.
}
\examples{
none
}
