% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDAL_resample_template.R
\name{gdal_resample_template}
\alias{gdal_resample_template}
\title{Resampling a Raster* object via GDAL}
\usage{
gdal_resample_template(r, r_base, method = "bilinear")
}
\arguments{
\item{r}{Raster* object to be resampled}

\item{r_base}{Raster* object with parameters that r should be resampled to.}

\item{method}{character. GDAL resampling_method ("near"|"bilinear"|"cubic"|"cubicspline"|
"lanczos"|"average"|"mode"|"max"|"min"|
"med"|"q1"|"q3")}
}
\description{
Resampling a Raster* object via GDAL
}
